Revert "Merge branch 'Calvin-Mongo'"
This reverts commit 79e8bdb3b862ea3e46b31642343ea80b69a2a352, reversing
changes made to fcdf89faa69fbe27e2de9c891a183ed5bb431413.

+++ src/main/java/base/studentFlowchartDb/Flowchart.java
@@ -13,13 +13,13 @@ public class Flowchart implements Serializable {
 
     @Id
     private String id;
-    private int studentId;
+    private Long studentId;
     private String flowchartName;
     private List<PlannedCourse> plannedCourses = new ArrayList<PlannedCourse>();
 
     public Flowchart() {}
 
-    public Flowchart(int studentId, String name, List<PlannedCourse> plannedCourses) {
+    public Flowchart(Long studentId, String name, List<PlannedCourse> plannedCourses) {
         this.studentId = studentId;
         this.flowchartName = name;
         this.plannedCourses.addAll(plannedCourses);
@@ -33,11 +33,11 @@ public class Flowchart implements Serializable {
         this.id = id;
     }
 
-    public int getStudentId() {
+    public Long getStudentId() {
         return studentId;
     }
 
-    public void setStudentId(int studentId) {
+    public void setStudentId(Long studentId) {
         this.studentId = studentId;
     }
